+++ Core/MyRect.cs
@@ -7,10 +7,10 @@ namespace DxfNet
 {
     public struct MyRect
     {
-        public int Left { get; set; }
-        public int Right { get; set; }
-        public int Top { get; set; }
-        public int Bottom { get; set; }
+        public float Left { get; set; }
+        public float Right { get; set; }
+        public float Top { get; set; }
+        public float Bottom { get; set; }
 
     }
 }
